backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
scikit-image
torch
transformers>=4.36
